Removed some boundary checking by mapping output to input matrix elements perfectly
The output matrix is smaller than the input matrix so as to avoid extrapolating element values that exceed image boundaries. In this commit, I removed the boundary checks for threads that apply the mask past the image boundary. And instead modified the CUDA kernel to perfectly map output elements to input elements. To make this work, I modified the CUDA grid dimensions to match the output matrix dimensions, rather than the input matrix dimensions.